requirements.txt
streamlit
pandas
openpyxl
python-docx
PyPDF2
Office365-REST-Python-Client
